backend-python/requirements.txt
flask
flask-cors
google-generativeai
python-docx
requests
groq
gunicorn
